Align sfm-sf06 driver package
The sfm-sf06 driver package was created one year ago and there where a
couple of small issues in it.
With the current update the driver package is aligned to the latest
python driver package style, the issues in the commands are fixed and
the latest libraries are referenced.
The tests and examples are executed again with an SFM4300.